split datasets and datasetapi
